Switch, alterações e testes.

+++ b4l/src/componentes/Card/index.js
@@ -15,41 +15,42 @@ import './index.css';
 
 export default function CustomCard(produto) {
   const classes = useStyles();
-  const { id, nome, descricao, preco, imagem } = produto.produto;
+  const { id, nome, descricao, preco, imagem, ativo } = produto.produto;
   const { token } = useAuth();
 
   const [erro, setErro] = useState('');
   const [carregando, setCarregando] = useState(false);
 
   async function handleExcluir() { 
-    setErro('');   
-    try {
-      const resposta = await fetch(`http://localhost:3000/produtos/${id}`, {
-        method: 'DELETE',
-        headers: {
-          'Authorization': `Bearer ${token}`
+    setErro('');
+    if (!ativo){
+      try {
+        const resposta = await fetch(`http://localhost:3000/produtos/${id}`, {
+          method: 'DELETE',
+          headers: {
+            'Authorization': `Bearer ${token}`
+          }
+        })
+        
+        const dados = await resposta.json();
+        
+        if (!resposta.ok) {
+          setErro(dados);
+          return;
         }
-      })
-      
-      const dados = await resposta.json();
-      
-      if (!resposta.ok) {
-        setErro(dados);
-        return;
+        produto.recarregar();
+      } catch (error) {
+        setErro(error.message)
       }
-      
-      produto.recarregar();
-    } catch (error) {
-      setErro(error.message)
-    }
+    }   
   };
   
   return (
-    <Card className={classes.root,  'cabelo'}>
+    <Card className={classes.root, 'cabelo'}>
       <div className={'eita'}>
         {erro && <Alert severity="error">{erro}</Alert>}
         <Button onClick={handleExcluir}>Excluir produto do catalogo</Button>
-        <CustomModal className='modal' acao='Editar produto' produtoInfo={produto.produto}/>
+        <CustomModal className='modal' acao='Editar produto' produtoInfo={produto.produto} recarregar={produto.recarregar} ativo={produto.ativo}/>
       </div>
       {carregando && <Loading/>}
       <CardActionArea className={classes.cardActionArea, 'adc-blur'}>
